# public/api/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vision Mission Career Portal API",
    "description": "Enterprise-grade, mobile-ready, versioned REST API architecture for the Career Portal. Supports JWT stateless authorization, strict parameter typing, rate limiting, and real-time document search.",
    "version": "1.0.0",
    "contact": {
      "name": "API Engineering Team",
      "email": "api-architect@visionmission.gov.in"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local environment"
    }
  ],
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "List Job Postings with Filtering & Paging",
        "description": "Returns a type-safe array of JobPost objects with paginator controls. Accepts query limits, search keywords, and location filters.",
        "parameters": [
          { "name": "search", "in": "query", "description": "Keyword search across titles and descriptions", "required": false, "schema": { "type": "string" } },
          { "name": "stateId", "in": "query", "description": "State identifier", "required": false, "schema": { "type": "integer" } },
          { "name": "categoryId", "in": "query", "description": "Job Category identifier", "required": false, "schema": { "type": "integer" } },
          { "name": "qualificationId", "in": "query", "description": "Required qualification identifier", "required": false, "schema": { "type": "integer" } },
          { "name": "minSalary", "in": "query", "description": "Minimum acceptable max salary range", "required": false, "schema": { "type": "number" } },
          { "name": "hasNoFee", "in": "query", "description": "Filter jobs requiring zero application fee", "required": false, "schema": { "type": "boolean" } },
          { "name": "postType", "in": "query", "description": "Categorized item type", "required": false, "schema": { "type": "string", "enum": ["job", "result", "admit_card", "syllabus", "notice"] } },
          { "name": "perPage", "in": "query", "description": "Items per page", "required": false, "schema": { "type": "integer", "default": 10 } },
          { "name": "page", "in": "query", "description": "Page offset identifier", "required": false, "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated listing retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Jobs list retrieved successfully." },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobs": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/JobPost" }
                        },
                        "pagination": { "$ref": "#/components/schemas/PaginationMetadata" }
                      }
                    },
                    "errors": { "type": "object", "nullable": true, "example": null },
                    "meta": { "$ref": "#/components/schemas/ResponseMeta" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{slug}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Retrieve Job Posting Detail",
        "description": "Fetches a full job detail representation matching the unique URL slug. Returns bookmarks flags dynamically if authenticated.",
        "parameters": [
          { "name": "slug", "in": "path", "description": "URL slug representing the job posting", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Detailed job parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Job details retrieved successfully." },
                    "data": { "$ref": "#/components/schemas/JobPost" },
                    "errors": { "type": "object", "nullable": true, "example": null },
                    "meta": { "$ref": "#/components/schemas/ResponseMeta" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Job posting not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Pass JWT access token under Authorization header in 'Bearer {token}' format"
      }
    },
    "schemas": {
      "ResponseMeta": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "integer", "example": 1716982200 },
          "api_version": { "type": "string", "example": "v1" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Validation failed." },
          "data": { "type": "object", "nullable": true, "example": null },
          "errors": {
            "type": "object",
            "nullable": true,
            "example": { "email": ["These credentials do not match our records."] }
          },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        }
      },
...
      "JobPost": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 101 },
          "postType": { "type": "string", "example": "job" },
          "title": { "type": "string", "example": "SSC CGL Inspector Recruitment" },
          "slug": { "type": "string", "example": "ssc-cgl-inspector-recruitment-8821" },
          "description": { "type": "string", "example": "<p>Central excise inspector openings available...</p>" },
          "examPattern": { "type": "string", "nullable": true, "example": "Written MCQ Exam Tiers" },
          "selectionProcess": { "type": "string", "nullable": true, "example": "Written + Interview" },
          "ageLimit": { "type": "string", "nullable": true, "example": "18-30 Years" },
          "salaryMin": { "type": "number", "nullable": true, "example": 44900.00 },
          "salaryMax": { "type": "number", "nullable": true, "example": 142400.00 },
          "vacancyCount": { "type": "integer", "example": 450 },
          "applicationFee": { "type": "number", "example": 100.00 },
          "officialWebsiteLink": { "type": "string", "format": "uri", "nullable": true, "example": "https://ssc.gov.in" },
          "applyLink": { "type": "string", "format": "uri", "nullable": true, "example": "https://ssc.gov.in/apply" },
          "notificationPdfPath": { "type": "string", "nullable": true, "example": "notifications/ssc_cgl_2026.pdf" },
          "lastDateToApply": { "type": "string", "format": "date", "nullable": true, "example": "2026-06-30" },
          "examDate": { "type": "string", "format": "date", "nullable": true, "example": "2026-08-15" },
          "publishedAt": { "type": "string", "format": "date-time", "nullable": true, "example": "2026-05-30T10:00:00+00:00" },
          "isFeatured": { "type": "boolean", "example": true },
          "category": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 3 },
              "name": { "type": "string", "example": "SSC" },
              "slug": { "type": "string", "example": "ssc" }
            }
          },
          "department": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 1 },
              "name": { "type": "string", "example": "Staff Selection Commission" }
            }
          },
          "state": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 10 },
              "name": { "type": "string", "example": "Central Government" }
            }
          },
          "district": { "type": "object", "nullable": true },
          "qualification": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 2 },
              "name": { "type": "string", "example": "Graduate" }
            }
          },
          "isBookmarked": { "type": "boolean", "example": false }
        }
      },
...
      "PaginationMetadata": {
        "type": "object",
        "properties": {
          "currentPage": { "type": "integer", "example": 1 },
          "totalPages": { "type": "integer", "example": 5 },
          "perPage": { "type": "integer", "example": 10 },
          "totalCount": { "type": "integer", "example": 48 },
          "hasNextPage": { "type": "boolean", "example": true },
          "hasPreviousPage": { "type": "boolean", "example": false }
        }
      }
    }
  }
}